requirements.txt
git+https://github.com/yaniv-aknin/fafdata
functions-framework
google-cloud-storage
